pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "otel-genai-graph"
version = "0.1.0"
description = "OpenTelemetry GenAI span → Neo4j graph exporter (v1.37 semconv)"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "Apache-2.0" }
authors = [{ name = "otel-genai-graph contributors" }]
keywords = ["opentelemetry", "genai", "neo4j", "observability", "llm", "graph"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Monitoring",
]
dependencies = [
    "opentelemetry-api>=1.27",
    "opentelemetry-sdk>=1.27",
    "neo4j>=5.20",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.6",
    "mypy>=1.11",
]
otlp = [
    "opentelemetry-exporter-otlp-proto-http>=1.27",
]

[project.urls]
Homepage = "https://github.com/kums/otel_genai_graph_exporter"
Issues = "https://github.com/kums/otel_genai_graph_exporter/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true
files = ["src"]
